+++ src/controllers/destacados.js
@@ -1,0 +1,18 @@
+const destacados =[
+    {
+        id: 1,
+        name: 'Consola',
+        precio: 100.0,
+        description: 'descripción no disponible',
+        img: '../img/Consola.jpg',
+    },
+    {
+        id: 5,
+        name: 'samsung-galaxy-s10',
+        precio: 500.0,
+        description: 'descripción no disponible',
+        img: '../img/img-samsung-galaxy-s10.jpg',
+    },
+];
+
+module.exports = destacados;
