+++ src/controllers/ofertas.js
@@ -1,0 +1,18 @@
+const ofertas =[
+    {
+        id: 2,
+        name: 'fpgaBoard',
+        precio: 200.0,
+        description: 'descripción no disponible',
+        img: '../img/fpgaBoard.jpg',
+    },
+    {
+        id: 3,
+        name: 'Libro',
+        precio: 300.0,
+        description: 'descripción no disponible',
+        img: '../img/image_book.jpg',
+    },
+];
+
+module.exports = ofertas;
