+++ src/routes/mainProducto.js
@@ -6,5 +6,4 @@ const router = express.Router();
 
 router.get('/:id', mainController.detalleProducto);
 
-
 module.exports = router;
