+++ src/views/partials/header.ejs
@@ -22,11 +22,11 @@
         <div class="nav-object">
             <ul class="nav">
                 <!-- <li class="nav-item"><a class="nav-link" href="detalleProducto">Detalle del producto</a></li> -->
-                <li class="nav-item"><a class="nav-link" href="categorias">Categorias</a></li>
-                <li class="nav-item"><a class="nav-link" href="crearLista">Crear lista</a></li>
-                <li class="nav-item"><a class="nav-link" href="registro">Registro</a></li>
-                <li class="nav-item"><a class="nav-link" href="login">login</a></li>
-                <li class="nav-item"><a class="nav-link" href="carritoCompras"><i class="fa-solid fa-cart-shopping"></i></a></li>
+                <li class="nav-item"><a class="nav-link" href="/categorias">Categorias</a></li>
+                <li class="nav-item"><a class="nav-link" href="/crearLista">Crear lista</a></li>
+                <li class="nav-item"><a class="nav-link" href="/registro">Registro</a></li>
+                <li class="nav-item"><a class="nav-link" href="/login">login</a></li>
+                <li class="nav-item"><a class="nav-link" href="/carritoCompras"><i class="fa-solid fa-cart-shopping"></i></a></li>
                 <li class="nav-item"><a class="nav-link" href=""><i class="fa-solid fa-bell"></i></a></li>
             </ul>
         </div>
